Fallando con unidad despues de quitar procesador. Cosas del ef core que aún no endiendo. Falla al modificar un workstation si selecionamos una persona distinta. (sí falla procesador y/o unidad)

+++ Inventario5/Models/Ordenador.cs
@@ -14,7 +14,7 @@ namespace Inventario5.Models
         public int? DiscoDuro { get; set; }
         public byte? UnidadDiscoDuro { get; set; }
         public int? IdProcesador { get; set; }
-        public int? VelocidadProcesador { get; set; }
+        public float? VelocidadProcesador { get; set; }
 
         public virtual ElementoInventariable IdNavigation { get; set; }
         public virtual Procesador IdProcesadorNavigation { get; set; }
